refactor(#9): presentation - service 간 통신 DTO 및 Mapper 클래스 제거
- Mapper 클래스 제거로 코드 단순화
- presentation과 service 간 공통 DTO 통합
