--- service/coupon-api/src/main/java/com/couponify/couponapi/application/dto/CouponCreateDto.java
@@ -1,16 +1,0 @@
-package com.couponify.couponapi.application.dto;
-
-import com.couponify.coupondomain.domain.coupon.CouponStatus;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CouponCreateDto {
-
-  private String name;
-  private CouponStatus status;
-  private Integer quantity;
-
-
-}
