--- service/coupon-api/src/main/java/com/couponify/couponapi/persentation/CouponControllerMapper.java
@@ -1,16 +1,0 @@
-package com.couponify.couponapi.persentation;
-
-import com.couponify.couponapi.application.dto.CouponCreateDto;
-import com.couponify.couponapi.persentation.request.CouponCreateRequest;
-
-public class CouponControllerMapper {
-
-  public static CouponCreateDto toCouponCreateDto(CouponCreateRequest couponCreateRequest) {
-    return new CouponCreateDto(
-        couponCreateRequest.getName(),
-        couponCreateRequest.getStatus(),
-        couponCreateRequest.getQuantity()
-    );
-  }
-
-}
